dynamic: Support reverse dynamic tracing
The reverse dynamic tracing is a technique to reduce tracing overhead
by disabling unwanted functions.  This is good when compiler doesn't
support the recent tracing options (-mnop-mcount) for some reason.

The normal dynamic tracing only enables functions match to given
pattern.  But this disables functions don't match to the pattern.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>