dynamic: Add -U option to exclude some functions
Check list of patterns for a symbol name and apply dynamic tracing
only if it survived after all the patterns.

The pattern list is maintained in the order it was given on the
command line (wrt -P and -U), so the last will win over the previous
ones.  Thus you need to be careful when specifying command line
options.  A general advice is you'd better having generic pattern
first and then appending specific case later.  For example, if you
want to apply it to all but 'foo' you need to do is:

  $ uftrace -P . -U foo  myprog   # ok

But if you do this with -U option first, the result will not be like
you expected:

  $ uftrace -U foo -P .  myprog   # ???

In this case, the later -P . matches to every name so the previous -U
option has no meaning.  So it'll apply dynamic tracing to all
functions regardless of the -U option.

I also added a logic that when only unpatch options are given, it'll
treat as if there's match-all option is given at the beginning.  So
the first example above can be written simply like:

  $ uftrace -U foo  myprog   # ok

Signed-off-by: Namhyung Kim <namhyung@gmail.com>