build: Detect python3 as well
As python2 is going to die, we need to convert script engine to use
python3 by default.  But it should support python2 too as a fallback.

This uses python-config script to detect build configuration for
python3 since I think it's a symlink to python3-config for systems
only have python3.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>